Rm outdated file and update import statements
