+++ server/src/mds3.init.js
@@ -31,7 +31,7 @@ import {
 import { get_samples, get_active_groupset } from './termdb.sql.js'
 import { server_init_db_queries } from './termdb.server.init.ts'
 import { barchart_data } from './termdb.barchart.js'
-import { mayInitiateScatterplots } from './termdb.scatter.js'
+import { mayInitiateScatterplots } from '../routes/termdb.sampleScatter.ts'
 import { mayInitiateMatrixplots, mayInitiateNumericDictionaryTermplots } from './termdb.matrix.js'
 import { add_bcf_variant_filter } from './termdb.snp.js'
 import { validate_correlationVolcano } from '#routes/correlationVolcano.ts'
